Have speaker independent of libsynth
Except for the use of ringbuf (independent of the rest of libsynth)
which is header-only so not a library linking dependency.
Specifically, speaker has no dependency on the unit generators.

Note that a friend function is used for the audio-playout callback in
order to not have this in public part of speaker (its concern is the
audio-playout backend, so it would be counter-intuitive to have the
callback function part of the public api).