Remove movement from synth/sizr
This construct is envelope-agnostic and calculates scaling
on ::y evaluation.  It is more efficient to perform stretch
on specific envelopes underlying data, like with a .stretch(w)
virtual member and omit any scaling on ::y evaluation in the case
of i.e punctual and tabular.  functional will still use a
stretched (scaling on evaluation).

When reading from i.e midi-file, the envelopes will have timed
data and it would be artificial to scale this to [0,1) just
to pass along the to-be-scaled factor and then performing the
multiplication during evaluation.

Also, the arguments for builders often had one, arbitrary chosen,
of the envelopes wrapped in a movement.  Now the envelopes are
stretched and the width is passed, in some cases no longer needed.