address hol blocking with a cv
the rendering with threads has a blocking problem when some pixels
are much heavier to generate than other ones, and the output ordering
buffer grows.  it was observed to grow so long that rendering seemed
to halt, and a pthread_yield call in this condition in practice was
a work around on my system.  in this commit this is replaced by
having threads wait on a condition variable when this happens, and
notify them when some output progress has occurred again.  note that
the head of line (hol) problem is still present, but the needed
avoidance of the wild output buffer growth once some blocking occurs
is now actually taken care of.